Added Cancel Key handling and removal of repeat downloads.

+++ IPSWdl/Program.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Net.Http;
+using System.Threading;
 using System.Threading.Tasks;
 using Mono.Options;
 using Utf8Json;
@@ -27,11 +28,12 @@ namespace IPSWdl
             string pathToStoreFiles = null;
             string searchTerm = null;
             var showHelp = false;
+            var downloadAll = false;
 
             var p = new OptionSet() {
                 "Usage: ipswdl [OPTIONS]",
                 "Downloads Apple IPSW firmware files.",
-                "By default, the latest IPSW is downloaded for every device available.",
+                "The latest firmware is downloaded by default.",
                 "",
                 "Options:",
                 { "p|path=", "Specifies the {PATH} where files will be downloaded too. If not specified, the CWD is used.",
@@ -39,6 +41,10 @@ namespace IPSWdl
                 { "s|search=",
                     "Only downloads for devices matching the {TERM}",
                     s => searchTerm = s},
+                { "A|All",
+                    "Downloads the newest firmware for all devices.",
+                    a => downloadAll = a != null
+                },
                 { "h|help",  "show this message and exit",
                     v => showHelp = v != null },
             };
@@ -52,11 +58,12 @@ namespace IPSWdl
             catch (OptionException e)
             {
                 Console.WriteLine(e.Message);
-                Console.WriteLine("Try `ipswdl --help' for more information.");
+                Console.WriteLine("Try `ipswdl --help` for more information.");
                 return;
             }
 
-            if (showHelp)
+            //quit early if neither -A or -s are passed
+            if (showHelp || (!downloadAll && searchTerm is null))
             {
                 p.WriteOptionDescriptions(Console.Out);
                 return;
@@ -69,7 +76,7 @@ namespace IPSWdl
             Console.WriteLine($"Got {TotalCount} devices!");
 
             var startTime = DateTime.Now;
-            Console.WriteLine("Starting Downloads... This may take a very long time. Any files still downloading when the program is closed will be corrupt.");
+            Console.WriteLine("Starting Downloads... This may take a very long time.");
 
             //Start a download of most recent IPSW for each device in serial. Downloading in parallel is better only in very high bandwidth scenarios. 
             if (searchTerm is null)
@@ -122,18 +129,65 @@ namespace IPSWdl
 
             Console.WriteLine($"Beginning to download {firmwareListing.name} {firmwareListing.firmwares[0].version}");
 
+            //If file has already been downloaded, skip
+            if (File.Exists(Path.Join(basePathToFolder, $@"/IPSW/{firmwareListing.name}/{firmwareListing.firmwares[0].version}.ipsw")))
+            {
+                ++TotalDone;
+                Console.Write($"{firmwareListing.name} {firmwareListing.firmwares[0].version} already exists. Skipping download");
+                Console.ForegroundColor = ConsoleColor.Red;
+                Console.WriteLine($"                      {(TotalDone / TotalCount) * 100}% complete");
+                Console.ForegroundColor = ConsoleColor.White;
+                return;
+            }
+
+            //Create file
             Directory.CreateDirectory(Path.Join(basePathToFolder, $@"/IPSW/{firmwareListing.name}/"));
             await using var fileStream = File.Create(Path.Join(basePathToFolder, $@"/IPSW/{firmwareListing.name}/{firmwareListing.firmwares[0].version}.ipsw"));
             await using var dlStream = await Client.GetStreamAsync(urlToDownload);
 
+            using var cts = new CancellationTokenSource();
+
+            //Delete currently downloading files if program is cancelled
+            void DeleteCorruptFile(object sender, ConsoleCancelEventArgs args)
+            {
+                cts.Cancel();
+                fileStream.Dispose();
+                dlStream.Dispose();
+
+                Console.ForegroundColor = ConsoleColor.Magenta;
+                Console.WriteLine("Stop Command recived. Deleting corrupt downloads...");
+                Console.ForegroundColor = ConsoleColor.Gray;
+                try
+                {
+                    File.Delete(Path.Join(basePathToFolder, $@"/IPSW/{firmwareListing.name}/{firmwareListing.firmwares[0].version}.ipsw"));
+                    Console.WriteLine($"Deleted {Path.Join(basePathToFolder, $@"/IPSW/{firmwareListing.name}/{firmwareListing.firmwares[0].version}.ipsw")}");
+                }
+                catch (Exception)
+                {
+                    Console.WriteLine($"There was an error deleting {Path.Join(basePathToFolder, $@"/IPSW/{firmwareListing.name}/{firmwareListing.firmwares[0].version}.ipsw")}");
+                }
+
+            }
+
+            Console.CancelKeyPress += DeleteCorruptFile;
+
             //download the file
-            await dlStream.CopyToAsync(fileStream);
+            try
+            {
+                await dlStream.CopyToAsync(fileStream, cts.Token);
+            }
+            catch (TaskCanceledException)
+            {
+                //ignore, handled by the delegate above.
+            }
+
+            Console.CancelKeyPress -= DeleteCorruptFile;
 
             ++TotalDone;
             Console.Write($"Finished downloading {firmwareListing.name} {firmwareListing.firmwares[0].version}");
             Console.ForegroundColor = ConsoleColor.Red;
             Console.WriteLine($"                      {(TotalDone/TotalCount) * 100}% complete");
-            Console.ForegroundColor = ConsoleColor.White;
+            Console.ForegroundColor = ConsoleColor.Gray;
         }
